requirements.txt
mysql-connector-python
mkdocs
mkdocs-macros-plugin